Reverted the extract_general.py script to the process used by extract_ala.py
I really wanted multiprocessing to speed this up, and I'll keep working on that for the glycine extract script, but it seems very complicated to do 5M operations efficiently without hitting memory errors -- either because I am saving too many frames to memory, or copies of the topology (require multiple copies to split the processing over multiple CPUs), or some other thing I'm not thinking of.

The approach used previously (which is still present in `extract_gly.py`) works for a small subset of the data to be extracted, but once I scale up to the full 5.5M row dataframe I hit a memory error.