Realized the issue with restart runs of the 22.8M simulation are due to my data file having the wrong masses for each atom.
Rather than referencing the masses at the top or in the second column of each atom index, there is an atom type index for each, and those indexes were deleted from the original header when I copied over Richard's header. So every mass in the simulation was wrong, leading to the wrong density.