Needed to remake the .data files used for lammps input
The masses in my data files were incorrect. Took a few tries / changes to make the correct pdb format, but `convert_dcd_to_data.py` should be used -- ONLY UNTIL THE XYZ IS CREATED

With the output .xyz file from that script, run `fix_pdb.py` to load the xyz into ASE and save it as a PDB -- this is the format needed to use Richard's `pdb2lmp.py` script.

With that pdb2lmp script, the correct format for an input file should be created. This .data file format is what is needed for restarting the simulation with the correct masses, etc.