Updated `dcd2data.py` with 'flush=True' on print statements so that I can see output in .out file produced.
This script loads the topology only once and iterates over all frame directories in a parent dir, skipping subdirs if a .data file already exists there. Once this finishes, I can do the same batch submission used on the 15 frames of interest.

The changes to `dcd2data.py` reduced the time to create a .data file from ~400 sec to ~180 sec, so this was a worthwhile change. Hopefully this finishes asap so I can start automating the quench of these 400 frames.